+++ contracts/mocks/ERC721PausableMock.sol
@@ -7,6 +7,8 @@ import "../token/ERC721/ERC721Pausable.sol";
  * This mock just provides a public mint, burn and exists functions for testing purposes
  */
 contract ERC721PausableMock is ERC721Pausable {
+    constructor (string memory name, string memory symbol) public ERC721(name, symbol) { }
+
     function mint(address to, uint256 tokenId) public {
         super._mint(to, tokenId);
     }
