+++ contracts/token/ERC721/ERC721Burnable.sol
@@ -7,7 +7,7 @@ import "./ERC721.sol";
  * @title ERC721 Burnable Token
  * @dev ERC721 Token that can be irreversibly burned (destroyed).
  */
-contract ERC721Burnable is Context, ERC721 {
+abstract contract ERC721Burnable is Context, ERC721 {
     /**
      * @dev Burns a specific ERC721 token.
      * @param tokenId uint256 id of the ERC721 token to be burned.
